requirements.txt
# ============================================================
# DEPENDENCIAS PRINCIPALES
# Tech Lead Demo Project - Django con arquitectura hexagonal
# ============================================================

# --- Core Framework ---
Django==4.2.9
djangorestframework==3.14.0
django-environ==0.11.2
django-cors-headers==4.3.1

# --- Base de datos SQL (PostgreSQL) ---
psycopg2-binary==2.9.9
django-extensions==3.2.3

# --- Base de datos NoSQL (MongoDB) ---
mongoengine==0.27.0
pymongo==4.6.1

# --- Redis (cache + sesiones + rate limiting) ---
redis==5.0.1
django-redis==5.4.0
celery==5.3.6

# --- Message Brokers ---
confluent-kafka==2.3.0          # Kafka
pika==1.3.2                     # RabbitMQ

# --- Seguridad y Auth ---
djangorestframework-simplejwt==5.3.1
django-guardian==2.4.0
cryptography==42.0.0

# --- Logging estructurado ---
structlog==24.1.0
sentry-sdk==1.40.0

# --- Metricas Prometheus ---
django-prometheus==2.3.1
prometheus-client==0.19.0

# --- Health Checks ---
django-health-check==3.18.1

# --- HTTP Client ---
httpx==0.26.0
requests==2.31.0

# --- Validacion dominio ---
pydantic==2.6.0

# --- Testing ---
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
faker==22.6.0

# --- Documentacion API ---
drf-spectacular==0.27.1
